add ResolvedRaw to config
